My first version of the program.
For the initial version of the program, which is not yet completed, I currently have only set up the inputs with their corresponding while loops to ensure users cannot enter text where numbers are required. Additionally, I have defined a function to clear the command prompt at certain points in the program.