Second Version of the program
Continuing with the development of the initial version of the programme, the next step involves initialising the camera. It is at this juncture that we specify the dimensions of the photograph, that is to say, the height and the width. These parameters have already been requested from the user previously, ensuring a personalised and tailored experience to their requirements. This stage is critical, as it readies the system to capture images with the desired specifications, reflecting the importance of properly gathering user inputs in the early phases of the programme.