My fourth version of the script
I had to make quite a few changes to the program. Although its main functionality remains the same, I needed to change some requirements, such as creating a new folder for every 1500 logs to save the photos. For this, I had to create some new functions. The first one, represented here, already existed, but the other two did not. The one to count the lines in the log file and the second one basically creates the new folder. Additionally, throughout the code, I have added various logging statements that will be stored in the log file. I encountered some issues as every time I executed the script, it created a new folder instead of saving the logs in the same one until there were 1500 logs, so I had to change part of the code.